Modularize dashboard code-behind
