+++ DivAcerManagerMax/Dashboard.Charts.cs
@@ -1,0 +1,85 @@
+using System;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.ComponentModel;
+using System.Diagnostics;
+using System.IO;
+using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Text.RegularExpressions;
+using System.Threading;
+using System.Threading.Tasks;
+using Avalonia.Animation;
+using Avalonia.Controls;
+using Avalonia.Media;
+using Avalonia.Styling;
+using Avalonia.Threading;
+using LiveChartsCore;
+using LiveChartsCore.SkiaSharpView;
+using LiveChartsCore.SkiaSharpView.Avalonia;
+using LiveChartsCore.SkiaSharpView.Painting;
+using Material.Icons.Avalonia;
+using SkiaSharp;
+
+namespace DivAcerManagerMax;
+
+public partial class Dashboard
+{
+    private void InitializeTemperatureGraph()
+    {
+        // Initialize collections
+        _cpuTempHistory = new ObservableCollection<double>();
+        _gpuTempHistory = new ObservableCollection<double>();
+
+        // Initialize series
+        _tempSeries = new ObservableCollection<ISeries>
+        {
+            new LineSeries<double>
+            {
+                Values = _cpuTempHistory,
+                Name = "CPU Temperature",
+                Stroke = new SolidColorPaint(SKColors.CornflowerBlue) { StrokeThickness = 3 },
+                GeometryStroke = new SolidColorPaint(SKColors.DeepSkyBlue),
+                GeometryFill = new SolidColorPaint(SKColors.DeepSkyBlue),
+                Fill = new SolidColorPaint(SKColors.Transparent),
+                GeometrySize = 5,
+                XToolTipLabelFormatter = chartPoint => $"CPU: {chartPoint.Label}°C"
+            },
+            new LineSeries<double>
+            {
+                Values = _gpuTempHistory,
+                Name = "GPU Temperature",
+                Stroke = new SolidColorPaint(SKColors.LimeGreen) { StrokeThickness = 3 },
+                GeometryFill = new SolidColorPaint(SKColors.GreenYellow),
+                GeometryStroke = new SolidColorPaint(SKColors.GreenYellow),
+                Fill = new SolidColorPaint(SKColors.Transparent),
+                GeometrySize = 5,
+                XToolTipLabelFormatter = chartPoint => $"GPU: {chartPoint.Label}°C"
+            }
+        };
+
+        // Initialize and configure the chart
+        _temperatureChart = this.FindControl<CartesianChart>("TemperatureChart");
+        if (_temperatureChart != null)
+        {
+            _temperatureChart.Series = _tempSeries;
+            _temperatureChart.XAxes = new List<Axis>
+            {
+                new()
+                {
+                    Name = "Time",
+                    IsVisible = false
+                }
+            };
+            _temperatureChart.YAxes = new List<Axis>
+            {
+                new()
+                {
+                    Name = "Temperature (°C)",
+                    NamePaint = new SolidColorPaint(SKColors.Gray),
+                    LabelsPaint = new SolidColorPaint(SKColors.Gray)
+                }
+            };
+        }
+    }
+}
